fix: model update

+++ emberlog/transcriber/whisper_fast.py
@@ -46,7 +46,7 @@ class WhisperConfig:
     model_name: str = os.getenv("WHISPER_MODEL", "large-v3")
     device: str = os.getenv("WHISPER_DEVICE", "cuda")  # "cuda" | "cpu"
     compute_type: str = os.getenv(
-        "WHISPER_COMPUTE_TYPE", "float32"
+        "WHISPER_COMPUTE_TYPE", "float16"
     )  # "float16" on GPU, "int8" or "float32" as needed
     vad_filter: bool = _bool_env("WHISPER_VAD_FILTER", True)
     vad_parameters: dict[str, object] = field(default_factory=_load_vad_params)
